requirements.txt
flask
werkzeug
torch
torchvision
timm
xgboost
joblib
shap
pillow
numpy
